requirements.txt
Django==1.6.2
South==0.8.4
beautifulsoup4==4.3.2
birdy==0.2
boto==2.26.0
cssselect==0.9.1
dj-database-url==0.2.2
dj-static==0.0.5
django-haystack==2.1.0
django-pylibmc==0.5.0
django-storages==1.1.8
django-tastypie==0.11.0
django-toolbelt==0.0.1
elasticsearch==1.0.0
gunicorn==18.0
lxml==3.3.2
newrelic==2.14.0.11
oauthlib==0.6.1
psycopg2==2.5.2
pyelasticsearch==0.6.1
pylibmc==1.2.3
pyquery==1.2.8
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2013.9
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.3.3
six==1.5.2
static==1.0.2
urllib3==1.7.1
wsgiref==0.1.2
